net-p2p/c-lightning: remove default_configdir patch
Per a discussion[1] with Luke-Jr, lightningd is intended to be run in the context of human user logins. Running it as a system service is not the designed use case. Therefore, lightning-cli must default to looking for the RPC socket in ~/.lightning, which users may symlink to the system-wide location /var/lib/lightning if they wish to make use of the system service.

[1]: https://gitlab.com/bitcoin/gentoo/merge_requests/18