net-p2p/c-lightning: depend on net-libs/libwally-core[elements]
Sadly, libwally-core has a different ABI depending on whether it was built with Elements support. C-Lightning requires Elements support in libwally-core but has no way of detecting at run-time when it's absent. This leads to segfaults. This commit adds the requirement that libwally-core was built with Elements support, thus resolving the segfaults in C-Lightning.