app-misc/moneychanger-9999: bump to EAPI=6 to resolve error
qmake-utils.eclass was complaining of "unsupported EAPI=5" for this ebuild.

I did not test the modified ebuild, but I did review the differences from EAPI=5 to EAPI=6 and visually checked the ebuild for constructs that could be affected. The get_libdir function is now part of EAPI=6, so inheriting the multilib eclass may no longer be necessary, but I left it in place to be safe.