net-p2p/pushpool-9999: update to EAPI=4 (for compatibility with git-r3.eclass)
I performed a visual check that none of the changes in EAPI=4 should cause problems in this ebuild, but I did not test it.

I went to EAPI=4 as the least invasive compliant change possible, but EAPI=4 is nonetheless ancient and deprecated. The maintainer of this ebuild ought to update it to use the latest EAPI.