net-p2p/c-lightning: support COMPAT flags with two-digit minor versions
Also, fix an off-by-one error in the flag defaulting logic: the
COMPAT_Vxyz flags are supposed to mean "maintain compatibility with
version x.y.z", so we should not disable the flag by default if the
user currently *has* version x.y.z installed.

The COMPAT_V090 flag was introduced in release 0.9.0 apparently by
mistake, as the change it guards should have been guarded by the
COMPAT_V082 flag that had already been introduced in that release
cycle. This is what led to the off-by-one error. All other COMPAT flags
were added in releases whose version numbers were strictly greater than
the version number embedded in the flag name, and that rule is being
upheld by the COMPAT_V0100 flag that is being introduced in release
0.10.1. Anyway, the improved logic will do the right thing regardless
of which of these two methodologies is followed by future COMPAT flags.