net-p2p/c-lightning: modify runscript to "need bitcoind" if local
Previously the OpenRC runscript had only "use bitcoind", but that
allowed the bitcoind service to be stopped before the lightningd
service had stopped. This commit improves the runscript so that the
lightningd service has a hard dependency on the bitcoind service if the
bitcoin-rpcconnect configuration parameter is missing or is set to
localhost, 127.0.0.1, or ::1.