dev-python/coincurve: less hackish hack for running tests
The issue is that Python searches ${S} first for the 'coincurve' module
directory and finds it there but then fails to find the native library
there since the native library is built to ${BUILD_DIR}/lib/coincurve.
Moving ${S}/coincurve out of the way temporarily while running the
tests allows Python to find the 'coincurve' module directory beneath
the build directory, and it does find the native library there.