dev-python/txtorcon: add all dev-python/* deps to BDEPEND when USE="doc"
Sphinx apparently needs to import txtorcon in order to generate its
documentation, so this necessitates having all the required dev-python/*
dependencies present on the build system. We don't need sys-process/
lsof, though, as that's only used when txtorcon is actually run. (It
*is* needed on the build system if tests are enabled.)

The slightly elaborate dependency specifications herein are because only
one Python implementation needs the complete set of deps on the build
system in order to build the documentation, but all enabled Python
implementations need all the deps on the runtime system.