move dev-python/mrkd -> app-text/mrkd
Mrkd is not a Python library, and the fact that it's implemented in
Python is irrelevant to its purpose. It is a text processing tool and
rightly belongs in the app-text category. Further, nothing is gained by
installing it for multiple Python implementations since its only use is
to be called via a scripted entry point, so switch it to single-impl.
Reverse dependencies should depend on it without any USE-dep specifier
since they'll be invoking the mrkd script that will always invoke the
correct (and only) interpreter regardless of the value of EPYTHON in the
calling ebuild environment.