net-p2p/core-lightning-0.11.2-r100: add unkeyworded backports revision
This revision applies a fairly hefty patchset backported from v0.12.0. It
is intended to include only the bugfixes and optimizations released in
v0.12.0 but without including any of the new features. Only trivial touch-
ups were made to the cherry-picked commits versus the upstream commits, but
because this patchset is **UNOFFICIAL**, it is released unkeyworded and
with no warranties whatsoever.

!! USE AT YOUR OWN RISK !!

See: https://github.com/ElementsProject/lightning/compare/v0.11.2...whitslack:lightning:0.11.2/backports