net-p2p/core-lightning-0.11.2-r106: revbump with more upstream patches
This backports project has expanded in scope to encompass some new
features from upstream, though it will not introduce any database schema
changes, so it should always be safe to downgrade to the released
version after trying this version.

The ebuild now contains a complete list of upstream commits between
v0.11.2 and the head of the master branch as of this writing. The
commits that require database schema changes or that dispense with
deprecated functionality are commented out. Feel free to use this ebuild
as a starting point for your own custom backports build, but be warned
that upstream provides no way to undo database migrations, so if you
start to wander down that path, you will not be able to wander back.