net-p2p/core-lightning: begin new 22.11.1 backports series
As a reminder, the ethos of these backports series is to incorporate
bugfixes and new features from upstream except for any that introduce
database schema changes. It should always be safe, therefore, to
downgrade to the same-numbered release version after trying a backports
version.

The backports ebuilds contain a complete list of upstream commits
between the base release and the head of the master branch. The commits
that require database schema changes or that dispense with deprecated
functionality are commented out. Feel free to use these ebuilds as a
starting point for your own custom backports build, but be warned that
upstream provides no way to undo database migrations, so if you start to
wander down that path, you will not be able to wander back.