net-p2p/core-lightning: revbump backports
With this bump comes a major shift in the ethos of these backports.
Previously they were constructed by starting at an old release and
cherry-picking commits from upstream to backport features and fixes,
resolving conflicts along the way. Now they are constructed by starting
at an old release and *merging* commits from upstream. Commits that are
to be skipped over (because they introduce or depend on database schema
changes) are reverted, and the reversion is merged. More information
about the workflow and how to audit it can be found at the link below.

See: https://github.com/whitslack/lightning/blob/info/README.md