net-p2p/core-lightning: suppress spurious rebuilds when USE="-man"
If app-text/lowdown is not installed, then the upstream configurator
sets HAVE_LOWDOWN to 0, which causes external/lowdown/lowdown.h to be
added to EXTERNAL_HEADERS, which are specified as prerequisites of
every object file. Thus, because such header file is absent, Make
always determines that every required object file needs rebuilt. This
commit adds a hack to eliminate the mention of this header file since
we only ever use the system-installed Lowdown or none at all.