net-p2p/bitcoin-core-25.0: introduce combined package
This package combines:
 - dev-util/bitcoin-tx (always installed)
 - net-p2p/bitcoin-cli (installed if USE="bitcoin-cli")
 - net-p2p/bitcoind    (installed if USE="daemon")
 - net-p2p/bitcoin-qt  (installed if USE="qt5")

To ease the transition, new ebuilds of the above packages at version
25.0 are added. These are empty packages that simply RDEPEND upon
net-p2p/bitcoin-core with the relevant USE flags.

Two new USE flags are introduced to control whether to embed internal
copies of LevelDB and libsecp256k1 into Bitcoin Core or to dynamically
link with the system-installed libraries.

The "sqlite" USE flag is now enabled by default, as descriptor wallets
are no longer experimental and are indeed now the default.

The "wallet" USE flag has been dropped, as it was redundant. If "berkdb"
and/or "sqlite" is enabled, then you get wallet support. If neither is
enabled, then you don't.

When a wallet is enabled, the 'bitcoin-wallet' utility is now installed.

The OpenRC init script has been overhauled to use -daemonwait so that
startup of other services needing bitcoind will be deferred until after
bitcoind is ready to accept RPCs.

The datadir that has long been located at the awkward
/var/lib/bitcoin/.bitcoin/ is now migrated to /var/lib/bitcoind/. The
ebuild installs a symlink to maintain backward compatibility on systems
where the older path still exists.

The init scripts now specify the location for the debug log file,
defaulting to /var/log/bitcoind/debug.log, instead of allowing it to be
written by default to /var/lib/bitcoin/.bitcoin/debug.log. It's an
important distinction, as /var/log may be located on a different storage
volume than /var/lib.

See: https://bugs.gentoo.org/890261